Feat: merge Newgroup page
- 분리되어 있던 result 를 newgroup 페이지에 포함시킴
- Newgroup 폴더 구조 변경
